Create Advanced CFD Course content
In our treatment of the physics of fluid flows we have added a summary of basic ideas underpinning Direct Numerical Simulations and Numerical Analysis. This course thoroughly introduce the coding based methods for the numerical problems and differential equations. The course is divided into three parts, the first half covers basic theoretical and physical descriptions of Fundamental Equations and Modern CFD. The second half will cover the transformation of differential equations into the algebraic equations, differencing methods and discretization methods. The last part contain the numerical analysis in the form of MATLAB and C++ coding. Coding based problems are tackled well in this course, so that the student of different backgrounds will learn more about related domains.